fix : 홈 가이드북 목록조회 수정 (#72)

See merge request s07-webmobile1-sub2/S07P12D203!120

+++ Back/aejimeongji/src/main/java/com/ssafy/aejimeongji/domain/repository/customrepository/GuideBookRepositoryCustomImpl.java
@@ -32,7 +32,7 @@ public class GuideBookRepositoryCustomImpl implements GuideBookRepositoryCustom 
     }
 
     private BooleanExpression getWeightBetween(Double dogWeight) {
-        return dogWeight != null ? guideBook.weightMin.lt(dogWeight).and(guideBook.weightMax.goe(dogWeight)) : null;
+        return dogWeight != null ? guideBook.weightMin.doubleValue().lt(dogWeight).and(guideBook.weightMax.doubleValue().goe(dogWeight)) : null;
     }
 
     private BooleanExpression getAgeAndWeightEq(Integer dogMonth, Double dogWeight) {
